pppEmission: align source string symbol and null-compare in destructor (#4786)

+++ src/pppEmission.cpp
@@ -332,7 +332,7 @@ void pppDestructEmission(pppEmission* pppEmission_, pppEmissionUnkC* param_2) {
     *(u32*)(model + 0x104) = 0;
 
     _WaitDrawDone__8CGraphicFPci(&Graphic, s_pppEmission_cpp_801db7e8, 0x118);
-    if (state[0] != 0) {
+    if ((s32)state[0] != 0) {
         pppHeapUseRate__FPQ27CMemory6CStage((CMemory::CStage*)state[0]);
         state[0] = 0;
     }
